first draft of BAC and Grant post-processing on drinks - untested.
this adds two new tables to the schema: BAC and GrantCharges. BAC should be
self explanatory; GrantCharges is a list of volume deductions for each grant.
after a pour, one or more grant crant charges will be generated, corresponding
to the same number of grants among which the drink volume was divided.

(this change isn't really anything new; it was just functionality that i ripped
out in the switchover to sqlobject. this data used to live as columns in the
drink table)