chore: revise trailing slash handling
Prior url definitions installed routes with a `r".../?$"` path regex, meaning the same
view would be served regardless of trailing slash presence.

With the previous change, we use `path(..)` in most places, which provides a convenient
shorthand for declaring path parameters (`<int:id>` etc). However we can no longer
specify a regex in that shorthand.

This creates a new problem: Routes that formerly got a direct response, for example
`GET /api/events/`, now get a 404. This can break existing API clients, of which I'm
not sure how many there are, nor do I know their slash-or-not behaviors.

To get around this we install a new middleware, `PathRewriteMiddleware`, which rewrites
the internal path -- without redirecting -- when a trailing slash is found.
We only activate this for `/api/` routes since browsers will deal with 301s as needed
just fine.