core: refactor: remove `KegbotBackend`
This layer was introduced long ago to support a project that is
no more (multi-tenant backends). It was also woefully incomplete,
as some operations were implement by the "backend" but others
interacted with model instances directly.

Removing this layer to make the django codebase more conventional
and to have one less component to maintain.